Add sgqlc as an optional dependency
[sgqlc](https://pypi.org/project/sgqlc/) is a python package providing
a simple graphql client. It's an incredibly lightweight package. The
cool thing it can do is automatically generate a graphql schema json
object given a graphql server, and additionally can automatically build
the python client for querying the graphql server and build the python
objects to parse the graphql server response to. We'll use these two
features in a coming commit.